backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==1.3.5
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
